move verified domain loader tests
