+++ api-js/src/loaders/verified-domains/__tests__/load-verified-domain-by-key.test.js
@@ -3,10 +3,10 @@ const { DB_PASS: rootPass, DB_URL: url } = process.env
 const { ArangoTools, dbNameFromFile } = require('arango-tools')
 const { setupI18n } = require('@lingui/core')
 
-const englishMessages = require('../locale/en/messages')
-const frenchMessages = require('../locale/fr/messages')
-const { makeMigrations } = require('../../migrations')
-const { verifiedDomainLoaderByKey } = require('../loaders')
+const englishMessages = require('../../../locale/en/messages')
+const frenchMessages = require('../../../locale/fr/messages')
+const { makeMigrations } = require('../../../../migrations')
+const { verifiedDomainLoaderByKey } = require('../..')
 
 describe('given a verifiedDomainLoaderByKey dataloader', () => {
   let query, drop, truncate, migrate, collections, i18n, domain1, domain2, org
